hoverIntent jQuery Plug-in (refs Author:briancherne)
hoverIntent is a plug-in that attempts to determine the user's intent...
like a crystal ball, only with mouse movement! It works like (and was
derived from) jQuery's built-in hover. However, instead of immediately
calling the onMouseOver function, it waits until the user's mouse slows
down enough before making the call.